Archived old files and split project into multiple files
Took old files out and then split the code from DRSProcessing into multiple files
