+++ .gitignore
@@ -21,3 +21,10 @@ $RECYCLE.BIN/
 # =========================
 # Operating System Files
 # =========================
+Archived old files/DRS Processing (DESKTOP-EPMC6AU's conflicted copy 2019-12-19).ipynb
+Archived old files/DRS Processing.ipynb
+Archived old files/python/InitialCodeTest.py
+Archived old files/InitialCodeTest.txt
+Archived old files/python/SecondCodeVersion-PreSPACY.py
+Archived old files/SecondCodeVersion-PreSPACY.txt
+Archived old files/Synonyms.ipynb
