+++ .idea/workspace.xml
@@ -5,20 +5,31 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="92560cf3-bb85-4b4b-bc90-84e6d49a5286" name="Default Changelist" comment="">
-      <change afterPath="$PROJECT_DIR$/.idea/OntoGapDetection.iml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/.idea/dictionaries/Daniel.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/.idea/inspectionProfiles/profiles_settings.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/.idea/misc.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/.idea/modules.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/.idea/vcs.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/python/DRS Processing.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/OntoGapDetection.iml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/OntoGapDetection.iml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/DRS Processing (DESKTOP-EPMC6AU's conflicted copy 2019-12-19).ipynb" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/DRS Processing.ipynb" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/InitialCodeTest.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/SecondCodeVersion-PreSPACY.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Synonyms.ipynb" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/python/DRSGraph.graphml" beforeDir="false" afterPath="$PROJECT_DIR$/python/DRSGraph.graphml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/python/DRSProcessing.py" beforeDir="false" afterPath="$PROJECT_DIR$/python/DRSProcessing.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/python/InitialCodeTest.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/python/SecondCodeVersion-PreSPACY.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/python/jsonFile.txt" beforeDir="false" afterPath="$PROJECT_DIR$/python/jsonFile.txt" afterDir="false" />
     </list>
     <option name="EXCLUDED_CONVERTED_TO_IGNORED" value="true" />
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
     <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="FileTemplateManagerImpl">
+    <option name="RECENT_TEMPLATES">
+      <list>
+        <option value="Python Script" />
+      </list>
+    </option>
   </component>
   <component name="Git.Settings">
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
@@ -87,7 +98,9 @@
       <map>
         <entry key="MAIN">
           <value>
-            <State />
+            <State>
+              <option name="COLUMN_ORDER" />
+            </State>
           </value>
         </entry>
       </map>
