+++ python/DRSGraph.graphml
@@ -268,10 +268,10 @@
 <edge id="0" source="Action6" target="ModPP7">
   <data key="d1">isModifiedBy</data>
 </edge>
-<edge id="0" source="Action6" target="Action10">
+<edge id="0" source="Action6" target="Action9">
   <data key="d1">triggers</data>
 </edge>
-<edge id="0" source="Action6" target="Action9">
+<edge id="0" source="Action6" target="Action10">
   <data key="d1">triggers</data>
 </edge>
 <edge id="0" source="Action6" target="Property11">
