+++ python/DRSProcessing.py
@@ -3,416 +3,10 @@ import networkx
 from nltk.corpus import wordnet
 import requests
 import json
-import re
 
-
-# Create Graph
-def generateItemGraph(graphNumber):
-    itemGraph = networkx.MultiDiGraph()
-    itemGraph.add_node('Item' + str(graphNumber), value='')
-    itemGraph.add_node('ItemName' + str(graphNumber), value='')
-    itemGraph.add_node('ItemAffordance' + str(graphNumber), value='')
-    itemGraph.add_node('ItemDescription' + str(graphNumber), value='')
-    itemGraph.add_node('ItemRole' + str(graphNumber), value='')
-
-    itemGraph.add_edge('Item' + str(graphNumber), 'ItemName' + str(graphNumber), value='ItemHasName')
-    itemGraph.add_edge('Item' + str(graphNumber), 'ItemAffordance' + str(graphNumber), value='ItemHasAffordance')
-    itemGraph.add_edge('Item' + str(graphNumber), 'ItemDescription' + str(graphNumber), value='ItemHasDescription')
-    itemGraph.add_edge('Item' + str(graphNumber), 'ItemRole' + str(graphNumber), value='ItemHasRole')
-
-    return itemGraph
-
-
-def generatePropGraph(graphNumber):
-    propGraph = networkx.MultiDiGraph()
-    propGraph.add_node('Property' + str(graphNumber), value='')
-    propGraph.add_node('PropertyAdjective' + str(graphNumber), value='')
-    propGraph.add_node('PropertySecondaryObject' + str(graphNumber), value='')
-    propGraph.add_node('PropertyTertiaryObject' + str(graphNumber), value='')
-    propGraph.add_node('PropertyDegree' + str(graphNumber), value='')
-    propGraph.add_node('PropertyCompTarget' + str(graphNumber), value='')
-
-    propGraph.add_edge('Property' + str(graphNumber), 'PropertyAdjective' + str(graphNumber),
-                       value='PropertyHasAdjective')
-    propGraph.add_edge('Property' + str(graphNumber), 'PropertySecondaryObject' + str(graphNumber),
-                       value='PropertyHasSecondaryObject')
-    propGraph.add_edge('Property' + str(graphNumber), 'PropertyTertiaryObject' + str(graphNumber),
-                       value='PropertyHasTertiaryObject')
-    propGraph.add_edge('Property' + str(graphNumber), 'PropertyDegree' + str(graphNumber), value='PropertyHasDegree')
-    propGraph.add_edge('Property' + str(graphNumber), 'PropertyCompTarget' + str(graphNumber),
-                       value='PropertyHasCompTarget')
-
-    return propGraph
-
-
-def generateActionGraph(graphNumber):
-    actionGraph = networkx.MultiDiGraph()
-    actionGraph.add_node('Action' + str(graphNumber), value='')
-    actionGraph.add_node('ActionVerb' + str(graphNumber), value='')
-
-    actionGraph.add_edge('Action' + str(graphNumber), 'ActionVerb' + str(graphNumber), value='ActionHasVerb')
-
-    return actionGraph
-
-
-def generateModPPGraph(graphNumber):
-    modPPGraph = networkx.MultiDiGraph()
-    modPPGraph.add_node('ModPP' + str(graphNumber), value='')
-    modPPGraph.add_node('ModPPPrep' + str(graphNumber), value='')
-
-    modPPGraph.add_edge('ModPP' + str(graphNumber), 'ModPPPrep' + str(graphNumber), value='ModPPHasPrep')
-
-    return modPPGraph
-
-
-class ItemGraph(object):
-    # Constructor
-    def __init__(self, graphNumber):
-        self.graphNumber = graphNumber
-        if graphNumber is not None:
-            self.graph = generateItemGraph(self.graphNumber)
-        else:
-            self.graph = None
-
-    # Generic append method based on whatever target is passed in
-    def __append(self, target, newValue):
-        currentValue = self.graph.nodes(data=True)[target + str(self.graphNumber)]['value']
-        if currentValue == '':
-            updatedValue = newValue
-        else:
-            updatedValue = currentValue + '|' + newValue
-        self.graph.nodes(data=True)[target + str(self.graphNumber)]['value'] = updatedValue
-
-    # Generic replace method based on whatever target is passed in
-    def __replace(self, target, newValue):
-        self.graph.nodes(data=True)[target + str(self.graphNumber)]['value'] = newValue
-
-    # Append/replace methods for each node value in Item Graph
-    def appendItemValue(self, newValue):
-        self.__append('Item', newValue)
-
-    def replaceItemValue(self, newValue):
-        self.__replace('Item', newValue)
-
-    def appendItemName(self, newName):
-        self.__append('ItemName', newName)
-
-    def replaceItemName(self, newName):
-        self.__replace('ItemName', newName)
-
-    def appendItemAffordance(self, newAffordance):
-        self.__append('ItemAffordance', newAffordance)
-
-    def replaceItemAffordance(self, newAffordance):
-        self.__replace('ItemAffordance', newAffordance)
-
-    def appendItemDescription(self, newDescription):
-        self.__append('ItemDescription', newDescription)
-
-    def replaceItemDescription(self, newDescription):
-        self.__replace('ItemDescription', newDescription)
-
-    def appendItemRole(self, newRole):
-        self.__append('ItemRole', newRole)
-
-    def replaceItemRole(self, newRole):
-        self.__replace('ItemRole', newRole)
-
-    # Method to get the type of graph
-    # def getTypeOfNode(self, node):
-    #    print(node)
-    #    print(self.graph.nodes[node])
-    #    return self.graph.nodes[node]['value']
-
-    # Method to find a node containing a given value
-    def FindItemWithValue(self, valueToFind):
-        if self.graph is not None:
-            # iterate through all graph nodes
-            for node, values in self.graph.nodes.data():
-                # print(node, values)
-                # If the current Node's value = the value passed in
-                if values['value'] == valueToFind:
-                    return node
-        return None
-
-    # Methods to add different types of edges between nodes
-    def addGroupMembershipEdges(self, groupNode, memberNode):
-        self.graph.add_edge(memberNode, groupNode, value='IsMemberOf')
-        self.graph.add_edge(groupNode, memberNode, value='HasMember')
-
-    def addNodeEquivalencyEdges(self, firstNode, secondNode):
-        self.graph.add_edge(firstNode, secondNode, value='IsEquivalentTo')
-        self.graph.add_edge(secondNode, firstNode, value='IsEquivalentTo')
-
-    def addCompositionEdges(self, composedNode, partOfNode):
-        self.graph.add_edge(composedNode, partOfNode, value='HasA')
-        self.graph.add_edge(partOfNode, composedNode, value='IsPartOf')
-
-    def addPropertyEdge(self, objectNode, propertyNode):
-        self.graph.add_edge(objectNode, propertyNode, value='Is')
-
-    # Methods to add different types of edges between nodes
-    def addActionPerformerEdges(self, performerNode, actionNode):
-        self.graph.add_edge(performerNode, actionNode, value='Performs')
-        self.graph.add_edge(actionNode, performerNode, value='IsPerformedBy')
-
-    def addActionTargetEdges(self, actionNode, targetNode):
-        self.graph.add_edge(actionNode, targetNode, value='HasTarget')
-        self.graph.add_edge(targetNode, actionNode, value='IsTargetOf')
-
-    def addModifierVerbEdges(self, modifierNode, verbNode):
-        self.graph.add_edge(modifierNode, verbNode, value='ModifiesVerb')
-        self.graph.add_edge(verbNode, modifierNode, value='isModifiedBy')
-
-    def addModifierObjectEdges(self, modifierNode, objectNode):
-        self.graph.add_edge(modifierNode, objectNode, value='modifiesObject')
-        self.graph.add_edge(objectNode, modifierNode, value='isModifiedBy')
-
-    def addConditionalTriggerEdges(self, ifNodeValue, thenNodeValue):
-        ifNode = self.FindItemWithValue(ifNodeValue)
-        thenNode = self.FindItemWithValue(thenNodeValue)
-        # We only want to trigger actions, not statement
-        if ifNode is not None and thenNode is not None:
-            if 'Action' in thenNode:
-                self.graph.add_edge(ifNode, thenNode, value='triggers')
-                self.graph.add_edge(thenNode, ifNode, value='isTriggeredBy')
-
-    # Methods to replace values of specific nodes
-    def ReplaceItemAffordanceAtSpecificNode(self, nodeToAddAffordance, newAffordance):
-        node = self.FindItemWithValue(nodeToAddAffordance)
-        if node is not None:
-            edgesFromNode = self.graph.edges(node, data=True)
-            for startNode, endNode, edgeValues in edgesFromNode:
-                # If an edge has the value ItemHasName, then we want to modify the end node
-                if edgeValues['value'] == 'ItemHasAffordance':
-                    # Update graph with name
-                    self.graph.nodes(data=True)[endNode]['value'] = newAffordance
-                    return True
-        else:
-            print("No node with direct object reference as value found")
-            return False
-
-    # Methods to replace values of specific nodes
-    def AppendItemAffordanceAtSpecificNode(self, nodeToAddAffordance, newAffordance):
-        node = self.FindItemWithValue(nodeToAddAffordance)
-        if node is not None:
-            edgesFromNode = self.graph.edges(node, data=True)
-            for startNode, endNode, edgeValues in edgesFromNode:
-                # If an edge has the value ItemHasName, then we want to modify the end node
-                if edgeValues['value'] == 'ItemHasAffordance':
-                    # Update graph with name
-                    currentValue = self.graph.nodes(data=True)[endNode]['value']
-                    if currentValue == '':
-                        updatedValue = newAffordance
-                    else:
-                        updatedValue = currentValue + '|' + newAffordance
-                    self.graph.nodes(data=True)[endNode]['value'] = updatedValue
-                    return True
-        else:
-            print("No node with direct object reference as value found")
-            return False
-
-    # Methods to replace values of specific nodes
-    def AppendValueAtSpecificNode(self, nodeToAddValue, newValue):
-        # Update graph with name
-        currentValue = self.graph.nodes(data=True)[nodeToAddValue]['value']
-        if currentValue == '':
-            updatedValue = newValue
-        else:
-            updatedValue = currentValue + '|' + newValue
-        self.graph.nodes(data=True)[nodeToAddValue]['value'] = updatedValue
-        return True
-
-    def ReplaceItemNameAtSpecificNode(self, nodeToAddName, newName):
-        # Find Node
-        node = self.FindItemWithValue(nodeToAddName)
-        if node is not None:
-            # Get list of edges from the node
-            edgesFromNode = self.graph.edges(node, data=True)
-            for startNode, endNode, edgeValues in edgesFromNode:
-                # If an edge has the value ItemHasName, then we want to modify the end node
-                if edgeValues['value'] == 'ItemHasName':
-                    # Update graph with name
-                    self.graph.nodes(data=True)[endNode]['value'] = newName
-                    return True
-        else:
-            print("No node with direct object reference as value found")
-            return False
-
-
-class PropertyGraph(object):
-    # Constructor
-    def __init__(self, graphNumber):
-        self.graphNumber = graphNumber
-        if graphNumber is not None:
-            self.graph = generatePropGraph(self.graphNumber)
-        else:
-            self.graph = None
-
-    # Generic append method based on whatever target is passed in
-    def __append(self, target, newValue):
-        currentValue = self.graph.nodes(data=True)[target + str(self.graphNumber)]['value']
-        if currentValue == '':
-            updatedValue = newValue
-        else:
-            updatedValue = currentValue + '|' + newValue
-        self.graph.nodes(data=True)[target + str(self.graphNumber)]['value'] = updatedValue
-
-    # Generic replace method based on whatever target is passed in
-    def __replace(self, target, newValue):
-        self.graph.nodes(data=True)[target + str(self.graphNumber)]['value'] = newValue
-
-    # Append/replace methods for each node value in Property Graph
-    def appendPropValue(self, newValue):
-        self.__append('Property', newValue)
-
-    def replacePropValue(self, newValue):
-        self.__replace('Property', newValue)
-
-    def appendPropAdj(self, newAdjective):
-        self.__append('PropertyAdjective', newAdjective)
-
-    def replacePropAdj(self, newValue):
-        self.__replace('PropertyAdjective', newValue)
-
-    def appendPropSecObj(self, newSecondaryObject):
-        self.__append('PropertySecondaryObject', newSecondaryObject)
-
-    def replacePropSecObj(self, newSecondaryObject):
-        self.__replace('PropertySecondaryObject', newSecondaryObject)
-
-    def appendPropTertObj(self, newTertiaryObject):
-        self.__append('PropertyTertiaryObject', newTertiaryObject)
-
-    def replacePropTertObj(self, newTertiaryObject):
-        self.__replace('PropertyTertiaryObject', newTertiaryObject)
-
-    def appendPropDegree(self, newDegree):
-        self.__append('PropertyDegree', newDegree)
-
-    def replacePropDegree(self, newDegree):
-        self.__replace('PropertyDegree', newDegree)
-
-    def appendPropCompTarget(self, newCompTarget):
-        self.__append('PropertyCompTarget', newCompTarget)
-
-    def replacePropCompTarget(self, newCompTarget):
-        self.__replace('PropertyCompTarget', newCompTarget)
-
-    # Method to get the type of graph
-    # def getTypeOfNode(self, node):
-    #    return self.graph.nodes[node]['value']
-
-    # Method to find a node containing a given value
-    def FindPropertyWithValue(self, valueToFind):
-        if self.graph is not None:
-            # iterate through all graph nodes
-            for node, values in self.graph.nodes.data():
-                # If the current Node's value = the value passed in
-                if values['value'] == valueToFind:
-                    return node
-        return None
-
-
-class ActionGraph(object):
-    # Constructor
-    def __init__(self, graphNumber):
-        self.graphNumber = graphNumber
-        if graphNumber is not None:
-            self.graph = generateActionGraph(self.graphNumber)
-        else:
-            self.graph = None
-
-    # Generic append method based on whatever target is passed in
-    def __append(self, target, newValue):
-        currentValue = self.graph.nodes(data=True)[target + str(self.graphNumber)]['value']
-        if currentValue == '':
-            updatedValue = newValue
-        else:
-            updatedValue = currentValue + '|' + newValue
-        self.graph.nodes(data=True)[target + str(self.graphNumber)]['value'] = updatedValue
-
-    # Generic replace method based on whatever target is passed in
-    def __replace(self, target, newValue):
-        self.graph.nodes(data=True)[target + str(self.graphNumber)]['value'] = newValue
-
-    # Append/replace methods for each node value in Property Graph
-    def appendActionValue(self, newValue):
-        self.__append('Action', newValue)
-
-    def replaceActionValue(self, newValue):
-        self.__replace('Action', newValue)
-
-    def appendActionVerb(self, newVerb):
-        self.__append('ActionVerb', newVerb)
-
-    def replaceActionVerb(self, newValue):
-        self.__replace('ActionVerb', newValue)
-
-    # Method to get the type of graph
-    # def getTypeOfNode(self, node):
-    #    return self.graph.nodes[node]['value']
-
-    # Method to find a node containing a given value
-    def FindActionWithValue(self, valueToFind):
-        if self.graph is not None:
-            # iterate through all graph nodes
-            for node, values in self.graph.nodes.data():
-                # If the current Node's value = the value passed in
-                if values['value'] == valueToFind:
-                    return node
-        return None
-
-
-# Modifier_PP (adv will need a different graph)
-class ModifierPPGraph(object):
-    # Constructor
-    def __init__(self, graphNumber):
-        self.graphNumber = graphNumber
-        if graphNumber is not None:
-            self.graph = generateModPPGraph(self.graphNumber)
-        else:
-            self.graph = None
-
-    # Generic append method based on whatever target is passed in
-    def __append(self, target, newValue):
-        currentValue = self.graph.nodes(data=True)[target + str(self.graphNumber)]['value']
-        if currentValue == '':
-            updatedValue = newValue
-        else:
-            updatedValue = currentValue + '|' + newValue
-        self.graph.nodes(data=True)[target + str(self.graphNumber)]['value'] = updatedValue
-
-    # Generic replace method based on whatever target is passed in
-    def __replace(self, target, newValue):
-        self.graph.nodes(data=True)[target + str(self.graphNumber)]['value'] = newValue
-
-    # Append/replace methods for each node value in Property Graph
-    def appendModPPValue(self, newValue):
-        self.__append('ModPP', newValue)
-
-    def replaceModPPValue(self, newValue):
-        self.__replace('ModPP', newValue)
-
-    def appendModPPPrep(self, newPreposition):
-        self.__append('ModPPPrep', newPreposition)
-
-    def replaceModPPPrep(self, newPreposition):
-        self.__replace('ModPPPrep', newPreposition)
-
-    # Method to get the type of graph
-    # def getTypeOfNode(self, node):
-    #    return self.graph.nodes[node]['value']
-
-    # Method to find a node containing a given value
-    def FindModWithValue(self, valueToFind):
-        if self.graph is not None:
-            # iterate through all graph nodes
-            for node, values in self.graph.nodes.data():
-                # If the current Node's value = the value passed in
-                if values['value'] == valueToFind:
-                    return node
-        return None
+from GraphGeneration import *
+from ConditionalHandling import *
+from LineCategorization import *
 
 
 class predicateSwitcher(object):
@@ -504,7 +98,7 @@ class predicateSwitcher(object):
             # Check if naming or setting an equivalency
             if 'named' in predSubjRef:
                 # If so call naming method
-                self.DRSGraph = nameItem(predSubjRef, predDirObjRef, self.DRSGraph)
+                self.DRSGraph = self.nameItem(predSubjRef, predDirObjRef, self.DRSGraph)
             # If not named(XYZ) but still has 4 components
             elif numberOfComponents == 4:
                 # Get nodes for both subject and direct object
@@ -685,6 +279,15 @@ class predicateSwitcher(object):
                 print(
                     "Error - Encountered duplicate reference for property but did not find adjective node to append to")
             return True
+
+    # CAN PROBABLY MAKE THE NAME CHANGE A CLASS FUNCTION ????  MAKE APPEND/REPLACE BE ABLE TO TAKE IN SPECIFIC TARGETS?
+    def nameItem(self, predSubjRef, predDirObjRef, DRSGraph):
+        # Get item name out of "named(XYZ)"
+        itemName = predSubjRef[predSubjRef.find("(") + 1:predSubjRef.find(")")]
+        # Replace the name
+        DRSGraph.ReplaceItemNameAtSpecificNode(predDirObjRef, itemName)
+        # Return graph
+        return DRSGraph
 
 
 # VERY INITIAL VERSION, MOST CODE COPIED FROM PREDICATESWITCHER
@@ -1215,120 +818,6 @@ def requestNewTermToNymCheck(originalTerm):
     return newTerm
 
 
-# CAN PROBABLY MAKE THE NAME CHANGE A CLASS FUNCTION ????  MAKE APPEND/REPLACE BE ABLE TO TAKE IN SPECIFIC TARGETS?
-def nameItem(predSubjRef, predDirObjRef, DRSGraph):
-    # Get item name out of "named(XYZ)"
-    itemName = predSubjRef[predSubjRef.find("(") + 1:predSubjRef.find(")")]
-    # Replace the name
-    DRSGraph.ReplaceItemNameAtSpecificNode(predDirObjRef, itemName)
-    # Return graph
-    return DRSGraph
-
-
-# Check if current line is a conditional or a header line
-def getSymbolLines(DRSLines):
-    # Declarations
-    symbolLines = {}
-    # Iterate through DRS commands
-    for index, line in enumerate(DRSLines):
-        # If line starts with bracket, it's a header
-        if line[0] == '[':
-            symbolLines.update({index: "header"})
-        # if line is arrow, it's a conditional
-        elif line == '=>':
-            symbolLines.update({index: "conditional"})
-        # if line is "QUESTION" then it's the start of a question segment
-        elif line == 'QUESTION':
-            symbolLines.update({index: "question-tag"})
-    print(symbolLines)
-    return symbolLines
-
-
-def categorizeSymbolLines(symbolLines):
-    # SymbolLines - check surrounding symbols and set instruction-header/if-header/then-header/question-header
-    # accordingly
-    symbolLineIndexes = list(symbolLines.keys())
-    categorizedSymbolLines = {}
-    for symbolIndex, symbolLineNumber in enumerate(symbolLineIndexes):
-        # Get symbol and surrounding symbols
-        currentSymbol = symbolLines[symbolLineNumber]
-        previousSymbol = None
-        nextSymbol = None
-        # If not first line, then set previous symbol
-        if symbolIndex != 0:
-            previousSymbolLineNumber = symbolLineIndexes[symbolIndex - 1]
-            previousSymbol = symbolLines[previousSymbolLineNumber]
-        # If not last line, then set next symbol
-        if symbolIndex < (len(symbolLines) - 1):
-            nextSymbolLineNumber = symbolLineIndexes[symbolIndex + 1]
-            nextSymbol = symbolLines[nextSymbolLineNumber]
-        # Categorize headers (ignore rest)
-        if currentSymbol == 'header':
-            # First header always = instruction as far as I've been able to see
-            if previousSymbol is None:
-                categorizedSymbolLines.update({symbolLineNumber: 'instruction-header'})
-            # If previous symbol is "QUESTION", then header is for a question
-            if previousSymbol == 'question-tag':
-                categorizedSymbolLines.update({symbolLineNumber: 'question-header'})
-            # if previous symbol is a conditional, then header is for a then part of the conditional
-            if previousSymbol == 'conditional':
-                categorizedSymbolLines.update({symbolLineNumber: 'then-header'})
-            # if next symbol is a conditional, then header is for an if part of the conditional
-            if nextSymbol == 'conditional':
-                categorizedSymbolLines.update({symbolLineNumber: 'if-header'})
-        else:
-            categorizedSymbolLines.update({symbolLineNumber: currentSymbol})
-    return categorizedSymbolLines
-
-
-# Categorize each variable in the DRS program with a type (instruction, if, then, question)
-def categorizeVariables(DRSLines, categorizedSymbolLines):
-    variablesAndTypes = {}
-    # Iterate through Symbol Lines in order to extract each variable from the headers and its associated type
-    for symbolLineNumber, symbolLineType in categorizedSymbolLines.items():
-        # check if header type
-        if 'header' in symbolLineType:
-            # get current line
-            currentSymbolLine = DRSLines[symbolLineNumber]
-            # strip brackets
-            currentSymbolLine = currentSymbolLine.replace('[', '')
-            currentSymbolLine = currentSymbolLine.replace(']', '')
-            # Get list of variables contained in header
-            currentTargetVariables = currentSymbolLine.split(',')
-            # Get current type of the header by stripping the header part of the tag and leaving the type
-            currentHeaderType = symbolLineType.split('-header')[0]
-            # Assign each variable in the current line to have the associated type with the header
-            for variable in currentTargetVariables:
-                variablesAndTypes.update({variable: currentHeaderType})
-    return variablesAndTypes
-
-
-# Assign a type (instruction, if, then, question) to each line of the DRS instructions in order to easily check
-# line-type during processing
-def categorizeDRSLines(DRSLines, symbolLines):
-    categorizedDRSLines = {}
-    # get categorization of each symbol line
-    categorizedSymbolLines = categorizeSymbolLines(symbolLines)
-    # Get categorizations of each variable
-    variablesAndTypes = categorizeVariables(DRSLines, categorizedSymbolLines)
-    # Iterate through DRSLines
-    for index, line in enumerate(DRSLines):
-        # If current line is in one of the symbol lines, give it the exact same type as it already has
-        if index in categorizedSymbolLines.keys():
-            currentLineType = categorizedSymbolLines.get(index)
-            categorizedDRSLines.update({index: currentLineType})
-        # otherwise, it's an instruction - check its reference variable and assign the associated type to the line
-        else:
-            # Get the contents of this given instruction
-            predicateContents = line.split('(', 1)[1]
-            # Get the reference variable of the instruction
-            predicateReferenceVariable = predicateContents.split(',', 1)[0]
-            # Get the type associated with this variable
-            currentInstructionType = variablesAndTypes.get(predicateReferenceVariable)
-            categorizedDRSLines.update({index: currentInstructionType})
-    return categorizedDRSLines
-
-
 def APEWebserviceCall(phraseToDRS):
     print(phraseToDRS)
     # Make APE Webservice call with given ACE phrase
@@ -1434,272 +923,6 @@ def getNyms(wordToCheck):
     # print(uniqueNymList)
     # print("ANTONYMS", uniqueAntonymList)
     return uniqueNymList, uniqueAntonymList
-
-
-class Conditional:
-    def __init__(self, firstLine):
-        self.firstLine = firstLine
-        self.anonFirstLine = ''
-        self.ifLines = []
-        self.anonIfLines = []
-        self.thenLines = []
-        self.anonThenLines = []
-        self.processed = False
-
-    def addIfLine(self, newIfLine):
-        self.ifLines.append(newIfLine)
-
-    def addThenLine(self, newThenLine):
-        self.thenLines.append(newThenLine)
-
-    def addAnonIfLine(self, newAnonIfLine):
-        self.anonIfLines.append(newAnonIfLine)
-
-    def addAnonThenLine(self, newAnonThenLine):
-        self.anonThenLines.append(newAnonThenLine)
-
-    def replaceIfLines(self, newIfLines):
-        self.ifLines = newIfLines
-
-    def replaceAnonIfLines(self, newAnonIfLines):
-        self.anonIfLines = newAnonIfLines
-
-    def replaceFirstLine(self, newFirstLine):
-        self.firstLine = newFirstLine
-
-    def replaceAnonFirstLine(self, newAnonFirstLine):
-        self.anonFirstLine = newAnonFirstLine
-
-    def markProcessed(self):
-        self.processed = True
-
-    def pprint(self):
-        print("firstLine: " + self.firstLine)
-        print("anonfirstLine: " + self.anonFirstLine)
-        print("IF: ")
-        print(self.ifLines)
-        print("THEN: ")
-        print(self.thenLines)
-        print("ANON IF: ")
-        print(self.anonIfLines)
-        print("ANON THEN: ")
-        print(self.anonThenLines)
-        print("PROCESSED: ")
-        print(self.processed)
-
-
-# Iterate through all reference variable/replacement number pairs and perform the substitution,
-# then return the anonymized if line
-def replaceReferenceVariables(ifLine, referenceVariables):
-    for originalVariable, replacementNumber in referenceVariables.items():
-        ifLine = re.sub(originalVariable, str(replacementNumber), ifLine)
-    return ifLine
-
-
-# Make the predicate reference variables irrelevant so that checking if an incoming instruction is equivalent to
-# a conditional if is easier
-def anonymizeIfs(conditionalToAnonymize):
-    # Get list of reference variables and assign a number to each to track order,
-    # then replace the variable with the number
-    referenceVariables = {}
-    currentVarNumber = 0
-    anonymizedIfLines = []
-    # assign a number to each if line's reference var in each conditional
-    for ifLine in conditionalToAnonymize.ifLines:
-        predicateContents = ifLine.split('(', 1)[1]
-        predicateReferenceVariable = predicateContents.split(',', 1)[0]
-        # Make sure not to overwrite something already found (for instance, modifiers have the same ref var)
-        if predicateReferenceVariable not in referenceVariables.keys():
-            referenceVariables.update({predicateReferenceVariable: currentVarNumber})
-        currentVarNumber = currentVarNumber + 1
-    # Now iterate through ifline's and replace each instance of a reference var with the associated number
-    for ifLine in conditionalToAnonymize.ifLines:
-        anonIfLine = replaceReferenceVariables(ifLine, referenceVariables)
-        anonymizedIfLines.append(anonIfLine)
-    conditionalToAnonymize.replaceAnonIfLines(anonymizedIfLines)
-    conditionalToAnonymize.replaceAnonFirstLine(anonymizedIfLines[0])
-    return conditionalToAnonymize
-
-
-def getConditionals(DRSLines, categorizedDRSLines):
-    # Need to make groups of which parts are related
-    conditionalLines = {}
-    for currentLineNumber in categorizedDRSLines.keys():
-        # reset conditional flag
-        # conditional = ''
-        # Find out if the current line is part of a conditional or not
-        # conditional = isLineConditional(index, symbolLines)
-        # if the current line is part of an IF in a conditional
-        if categorizedDRSLines.get(currentLineNumber) == 'if':
-            conditionalLines.update({currentLineNumber: 'if'})
-
-        # if the current line is part of a THEN in a conditional
-        if categorizedDRSLines.get(currentLineNumber) == 'then':
-            conditionalLines.update({currentLineNumber: 'then'})
-
-    # print(conditionalLines)
-    # print(conditionalLines.values())
-
-    # iterate through and group each line in the same if/then, then match them up
-    # TODO: Make the string split to remove the numbers at the end of the DRS instruction happen when it's first read in
-    # so that it doesn't have to be done every time the instruction is touched in this or the switcher.
-    # Also this method is clunky and temporary.  WIll need cleaned up.
-    conditionalList = []
-    conditionalLineIndexes = list(conditionalLines.keys())
-    for conditionalIndex, conditionalLineNumber in enumerate(conditionalLineIndexes):
-        # print(conditionalIndex, conditionalLineNumber)
-        # If item is an if
-        if conditionalLines[conditionalLineNumber] == 'if':
-            # Get index for line
-            ifLineIndex = conditionalLineNumber
-            # If first conditional line overall, or first line since a then
-            if (conditionalIndex - 1) < 0 or conditionalLines[conditionalLineIndexes[conditionalIndex - 1]] == 'then':
-                # Create new Conditional with the current line as the first line in the conditional
-                currentConditional = Conditional(DRSLines[ifLineIndex].split(')-')[0] + ')')
-            # Otherwise, just an if line
-            currentConditional.addIfLine(DRSLines[ifLineIndex].split(')-')[0] + ')')
-        # If item is a then
-        elif conditionalLines[conditionalLineNumber] == 'then':
-            thenLineIndex = conditionalLineNumber
-            currentConditional.addThenLine(DRSLines[thenLineIndex].split(')-')[0] + ')')
-            # If last line overall or last then before an if
-            if ((conditionalIndex + 1) >= (len(conditionalLineIndexes))) or\
-                    conditionalLines[conditionalLineIndexes[conditionalIndex + 1]] == 'if':
-                currentConditional = anonymizeIfs(currentConditional)
-                conditionalList.append(currentConditional)
-    # COMMENTED OUT - we don't want to pre-anonymize, rather anonymize on check
-    # conditionalList = anonymizeIfs(conditionalList)
-    # for conditional in conditionalList:
-    #            conditional.pprint()
-    return conditionalList
-
-
-def splitAndRun(currentInstruction, predSwitcher):
-    predicateSplit = currentInstruction.split('(', 1)
-    predicateType = predicateSplit[0]
-    predicateContents = predicateSplit[1]
-    # Call appropriate handling function based on predicate type
-    DRSGraph = predSwitcher.callFunction(predicateType, predicateContents)
-    return DRSGraph
-
-
-# Removed possibly unnecessary parameter - predSwitcher
-def runConditionalThen(conditional, conditionalSets):
-    newDRSLines = []
-    # Run each then line in the triggered conditional, and add it to a "new DRS lines" to check for chain reaction
-    for triggeredThenLine in conditional.thenLines:
-        # Run the current line - may have been completely unnecessary
-        # DRSGraph = splitAndRun(triggeredThenLine, predSwitcher)
-        # Adding -0/0 to end of line to mimic numbers at end of normal line and not mess up splits
-        newDRSLines.append(triggeredThenLine + '-0/0')
-    # Recursively run this checker on the newly triggered lines to see if they trigger anything
-    # , predSwitcher, DRSGraph - Removed potentially unnecessary arguments
-    DRSGraph = checkCurrentInstructionIf(newDRSLines, 0, newDRSLines[0], conditionalSets)
-    return DRSGraph
-
-
-def runFullConditional(conditional, predSwitcher, DRSGraph, conditionalSets):
-    checkPreparedThenLines = []
-    newIfNodes = []
-    newThenNodes = []
-    instructionCountInMatchingIfBlock = 0
-
-    # HACKY WAY to avoid second parentheses being appended due to not having a -x/y at the end
-    for thenLine in conditional.thenLines:
-        checkPreparedThenLines.append(thenLine + '-0/0')
-    # Run each line of the if part of the conditional
-    for ifLine in conditional.ifLines:
-        # Append newly created node to array - should probably make this a separate  function
-        predicateSplit = ifLine.split('(', 1)
-        predicateContents = predicateSplit[1]
-        ifLineNodeReference = predicateContents.split(',')[0]
-        newIfNodes.append(ifLineNodeReference)
-        # Run the current line
-        DRSGraph = splitAndRun(ifLine, predSwitcher)
-    # Run each then line in the then part of the conditional
-    for index, thenLine in enumerate(checkPreparedThenLines):
-        # Run the current then line if it isn't in a matching if block
-        if instructionCountInMatchingIfBlock == 0:
-            # , predSwitcher, DRSGraph - removed potentially unneeded parameters
-            instructionCountInMatchingIfBlock, conditionalWithMatchingIfBlock = checkCurrentInstructionIf(
-                checkPreparedThenLines, index, thenLine, conditionalSets)
-            if instructionCountInMatchingIfBlock == 0:
-                # Append newly created node to array - should probably make this a separate  function
-                predicateSplit = thenLine.split('(', 1)
-                predicateContents = predicateSplit[1]
-                thenLineNodeReference = predicateContents.split(',')[0]
-                newThenNodes.append(thenLineNodeReference)
-                # Run the current line
-                DRSGraph = splitAndRun(thenLine, predSwitcher)
-            else:
-                # Get the if block that matches the then block and add it to newThenNodes, in order to still
-                # be able to put in the conditional trigger edges
-                conditional.pprint()
-                for i in range(instructionCountInMatchingIfBlock):
-                    currentIfLine = conditionalWithMatchingIfBlock.ifLines[i]
-                    predicateSplit = currentIfLine.split('(', 1)
-                    predicateContents = predicateSplit[1]
-                    currentIfLineNodeReference = predicateContents.split(',')[0]
-                    print("**************************************************", currentIfLineNodeReference)
-                    newThenNodes.append(currentIfLineNodeReference)
-        else:
-            instructionCountInMatchingIfBlock = instructionCountInMatchingIfBlock - 1
-    # make lists contain only distinct members
-    newIfNodes = list(set(newIfNodes))
-    newThenNodes = list(set(newThenNodes))
-    # Add edges between if and then nodes to signify which nodes get triggered by conditional
-    print("IF", newIfNodes)
-    print("THEN", newThenNodes)
-    for ifNode in newIfNodes:
-        for thenNode in newThenNodes:
-            if ifNode is not None and thenNode is not None:
-                DRSGraph.addConditionalTriggerEdges(ifNode, thenNode)
-    conditional.markProcessed()
-    return DRSGraph
-
-
-# Potentially unused parameters removed
-# , predSwitcher, DRSGraph)
-def checkCurrentInstructionIf(DRSLines, currentInstructionIndex, currentInstruction, conditionalSets):
-    # Format instruction in the format of the conditionalSet
-    currentInstruction = currentInstruction.split(')-')[0] + ')'
-    # ASSUMPTION: Should not see a first line of an if that references another reference variable from the conditional
-    # Hacky way to get current instruction anonymized - should rework this
-    initialConditionalCheck = Conditional(currentInstruction)
-    initialConditionalCheck.addIfLine(currentInstruction)
-    initialConditionalCheck = anonymizeIfs(initialConditionalCheck)
-    conditionalWithMatchingIfBlock = None
-
-    instructionCountInMatchingIfBlock = 0
-    for conditional in conditionalSets:
-        # In the case that the current line matches a conditional found
-        if conditional.anonFirstLine == initialConditionalCheck.anonFirstLine:
-            # Build out a "potential conditional" to see if it matches the conditional we expect
-            conditionalIfLength = len(conditional.ifLines)
-            potentialConditional = Conditional(currentInstruction)
-            for i in range(conditionalIfLength):
-                # add n if lines, where n is the number of if lines in the conditional being checked against
-                potentialIfLine = DRSLines[currentInstructionIndex + i].split(')-')[0] + ')'
-                potentialConditional.addIfLine(potentialIfLine)
-            # Anonymize the lines from the instructions to see if they match the anonymized if
-            # lines from the conditional
-            potentialConditional = anonymizeIfs(potentialConditional)
-            # compare instructions and conditional
-            # potentialConditional.pprint()
-            matchingLines = 0
-            # Iterate through all the if lines in the conditional - if they match, add to the counter
-            for i in range(conditionalIfLength):
-                if conditional.anonIfLines[i] == potentialConditional.anonIfLines[i]:
-                    matchingLines = matchingLines + 1
-            # If the number of matching lines is the number of lines in the conditional, the conditional triggers
-            if matchingLines == conditionalIfLength:
-                # Return how many lines match and which conditional matches
-                instructionCountInMatchingIfBlock = conditionalIfLength
-                conditionalWithMatchingIfBlock = conditional
-                # predSwitcher - removed potentially unnecessary parameter
-                # DRSGraph = runConditionalThen(conditional, conditionalSets)
-    # return DRSGraph
-    return instructionCountInMatchingIfBlock, conditionalWithMatchingIfBlock
 
 
 # TODO: handle 5-item predicate() tags
