+++ python/LineCategorization.py
@@ -1,0 +1,104 @@
+
+
+# Check if current line is a conditional or a header line
+def getSymbolLines(DRSLines):
+    # Declarations
+    symbolLines = {}
+    # Iterate through DRS commands
+    for index, line in enumerate(DRSLines):
+        # If line starts with bracket, it's a header
+        if line[0] == '[':
+            symbolLines.update({index: "header"})
+        # if line is arrow, it's a conditional
+        elif line == '=>':
+            symbolLines.update({index: "conditional"})
+        # if line is "QUESTION" then it's the start of a question segment
+        elif line == 'QUESTION':
+            symbolLines.update({index: "question-tag"})
+    print(symbolLines)
+    return symbolLines
+
+
+def categorizeSymbolLines(symbolLines):
+    # SymbolLines - check surrounding symbols and set instruction-header/if-header/then-header/question-header
+    # accordingly
+    symbolLineIndexes = list(symbolLines.keys())
+    categorizedSymbolLines = {}
+    for symbolIndex, symbolLineNumber in enumerate(symbolLineIndexes):
+        # Get symbol and surrounding symbols
+        currentSymbol = symbolLines[symbolLineNumber]
+        previousSymbol = None
+        nextSymbol = None
+        # If not first line, then set previous symbol
+        if symbolIndex != 0:
+            previousSymbolLineNumber = symbolLineIndexes[symbolIndex - 1]
+            previousSymbol = symbolLines[previousSymbolLineNumber]
+        # If not last line, then set next symbol
+        if symbolIndex < (len(symbolLines) - 1):
+            nextSymbolLineNumber = symbolLineIndexes[symbolIndex + 1]
+            nextSymbol = symbolLines[nextSymbolLineNumber]
+        # Categorize headers (ignore rest)
+        if currentSymbol == 'header':
+            # First header always = instruction as far as I've been able to see
+            if previousSymbol is None:
+                categorizedSymbolLines.update({symbolLineNumber: 'instruction-header'})
+            # If previous symbol is "QUESTION", then header is for a question
+            if previousSymbol == 'question-tag':
+                categorizedSymbolLines.update({symbolLineNumber: 'question-header'})
+            # if previous symbol is a conditional, then header is for a then part of the conditional
+            if previousSymbol == 'conditional':
+                categorizedSymbolLines.update({symbolLineNumber: 'then-header'})
+            # if next symbol is a conditional, then header is for an if part of the conditional
+            if nextSymbol == 'conditional':
+                categorizedSymbolLines.update({symbolLineNumber: 'if-header'})
+        else:
+            categorizedSymbolLines.update({symbolLineNumber: currentSymbol})
+    return categorizedSymbolLines
+
+
+# Categorize each variable in the DRS program with a type (instruction, if, then, question)
+def categorizeVariables(DRSLines, categorizedSymbolLines):
+    variablesAndTypes = {}
+    # Iterate through Symbol Lines in order to extract each variable from the headers and its associated type
+    for symbolLineNumber, symbolLineType in categorizedSymbolLines.items():
+        # check if header type
+        if 'header' in symbolLineType:
+            # get current line
+            currentSymbolLine = DRSLines[symbolLineNumber]
+            # strip brackets
+            currentSymbolLine = currentSymbolLine.replace('[', '')
+            currentSymbolLine = currentSymbolLine.replace(']', '')
+            # Get list of variables contained in header
+            currentTargetVariables = currentSymbolLine.split(',')
+            # Get current type of the header by stripping the header part of the tag and leaving the type
+            currentHeaderType = symbolLineType.split('-header')[0]
+            # Assign each variable in the current line to have the associated type with the header
+            for variable in currentTargetVariables:
+                variablesAndTypes.update({variable: currentHeaderType})
+    return variablesAndTypes
+
+
+# Assign a type (instruction, if, then, question) to each line of the DRS instructions in order to easily check
+# line-type during processing
+def categorizeDRSLines(DRSLines, symbolLines):
+    categorizedDRSLines = {}
+    # get categorization of each symbol line
+    categorizedSymbolLines = categorizeSymbolLines(symbolLines)
+    # Get categorizations of each variable
+    variablesAndTypes = categorizeVariables(DRSLines, categorizedSymbolLines)
+    # Iterate through DRSLines
+    for index, line in enumerate(DRSLines):
+        # If current line is in one of the symbol lines, give it the exact same type as it already has
+        if index in categorizedSymbolLines.keys():
+            currentLineType = categorizedSymbolLines.get(index)
+            categorizedDRSLines.update({index: currentLineType})
+        # otherwise, it's an instruction - check its reference variable and assign the associated type to the line
+        else:
+            # Get the contents of this given instruction
+            predicateContents = line.split('(', 1)[1]
+            # Get the reference variable of the instruction
+            predicateReferenceVariable = predicateContents.split(',', 1)[0]
+            # Get the type associated with this variable
+            currentInstructionType = variablesAndTypes.get(predicateReferenceVariable)
+            categorizedDRSLines.update({index: currentInstructionType})
+    return categorizedDRSLines
